+++ Glyssen/Dialogs/QuotationMarksDlg.cs
@@ -15,8 +15,6 @@ using GlyssenEngine.Paratext;
 using GlyssenEngine.Quote;
 using GlyssenEngine.Script;
 using L10NSharp;
-using L10NSharp.XLiffUtils;
-using L10NSharp.UI;
 using SIL.ObjectModel;
 using SIL.Scripture;
 using SIL.Windows.Forms.Extensions;
@@ -27,7 +25,7 @@ using BlockNavigatorViewModel = GlyssenEngine.ViewModels.BlockNavigatorViewModel
 
 namespace Glyssen.Dialogs
 {
-	public partial class QuotationMarksDlg : FormWithPersistedSettings
+	public partial class QuotationMarksDlg : FormWithPersistedSettings, ILocalizable
 	{
 		private readonly Project m_project;
 		private readonly BlockNavigatorViewModel m_navigatorViewModel;
@@ -64,7 +62,6 @@ namespace Glyssen.Dialogs
 				m_scriptureReference.VerseControl.ShowEmptyBooks = false;
 
 				m_scriptureReference.VerseControl.AllowVerseSegments = false;
-				m_scriptureReference.VerseControl.Versification = m_navigatorViewModel.Versification;
 				m_scriptureReference.VerseControl.VerseRefChanged += m_scriptureReference_VerseRefChanged;
 
 				m_blocksViewer.Initialize(m_navigatorViewModel);
@@ -89,7 +86,7 @@ namespace Glyssen.Dialogs
 			try
 			{
 				HandleStringsLocalized();
-				LocalizeItemDlg<XLiffDocument>.StringsLocalized += HandleStringsLocalized;
+				Program.RegisterLocalizable(this);
 
 				SetFilterControlsFromMode();
 
@@ -109,7 +106,7 @@ namespace Glyssen.Dialogs
 			LoadBlock();
 		}
 
-		private void HandleStringsLocalized()
+		public void HandleStringsLocalized()
 		{
 			L10N.LocalizeComboList(m_toolStripComboBoxFilter, "DialogBoxes.QuotationMarksDlg.FilterOptions");
 
@@ -764,10 +761,7 @@ namespace Glyssen.Dialogs
 		{
 			if (m_navigatorViewModel == null)
 				return;
-			var blockRef = m_navigatorViewModel.GetBlockVerseRef();
-			int versesInBlock = m_navigatorViewModel.CurrentBlock.LastVerseNum - blockRef.VerseNum;
-			var displayedRefMinusBlockStartRef = m_scriptureReference.VerseControl.VerseRef.BBBCCCVVV - blockRef.BBBCCCVVV;
-			if (displayedRefMinusBlockStartRef < 0 || displayedRefMinusBlockStartRef > versesInBlock)
+			if (m_navigatorViewModel.IsReferenceOutsideCurrentScope(m_scriptureReference.VerseControl.VerseRef))
 				m_scriptureReference.VerseControl.VerseRef = m_navigatorViewModel.GetBlockVerseRef();
 			m_labelXofY.Visible = m_navigatorViewModel.IsCurrentLocationRelevant;
 			Debug.Assert(m_navigatorViewModel.RelevantBlockCount >= m_navigatorViewModel.CurrentDisplayIndex);